Refactor package structure
- Move all apps into single `health_check` package
- None Django backends move in `health_check.contrib`
- Update requirements
