+++ MANIFEST.in
@@ -1,2 +1,2 @@
-recursive-include health_check/templates *
-include README.md
+include LICENSE
+recursive-include health_check *
