+++ health_check/cache/plugin_health_check.py
@@ -24,4 +24,5 @@ class CacheBackend(BaseHealthCheckBackend):
         except Exception:
             raise ServiceUnavailable("Unknown exception")
 
+
 plugin_dir.register(CacheBackend)
