+++ health_check/db/models.py
@@ -4,3 +4,6 @@ from django.db import models
 
 class TestModel(models.Model):
     title = models.CharField(max_length=128)
+
+    class Meta:
+        db_table = 'health_check_db_testmodel'
