+++ health_check/db/plugin_health_check.py
@@ -6,8 +6,8 @@ from django.db import DatabaseError, IntegrityError
 from health_check.backends.base import (
     BaseHealthCheckBackend, ServiceReturnedUnexpectedResult, ServiceUnavailable
 )
+from health_check.db.models import TestModel
 from health_check.plugins import plugin_dir
-from health_check_db.models import TestModel
 
 logger = logging.getLogger(__name__)
 
@@ -27,5 +27,6 @@ class DjangoDatabaseBackend(BaseHealthCheckBackend):
         except DatabaseError:
             logger.exception("Database Error")
             raise ServiceUnavailable("Database error")
+
 
 plugin_dir.register(DjangoDatabaseBackend)
