+++ health_check/storage/plugin_health_check.py
@@ -2,10 +2,11 @@
 from django.conf import settings
 
 from health_check.plugins import plugin_dir
-from health_check_storage.base import StorageHealthCheck
+from .base import StorageHealthCheck
 
 
 class DefaultFileStorageHealthCheck(StorageHealthCheck):
     storage = settings.DEFAULT_FILE_STORAGE
+
 
 plugin_dir.register(DefaultFileStorageHealthCheck)
