+++ requirements-dev.txt
@@ -1,14 +1,27 @@
--e .
-celery==4.0.1
-coverage==4.0.3
-django-coverage-plugin==1.3
-flake8==2.5.4
+#
+# This file is autogenerated by pip-compile
+# To update, run:
+#
+#    pip-compile --output-file requirements-dev.txt requirements-dev.in
+#
+amqp==2.1.4               # via kombu
+billiard==3.5.0.2         # via celery
+celery==4.0.2
+coverage==4.3.1
+flake8==3.2.1
 isort==4.2.5
-mccabe==0.4.0
+kombu==4.0.2              # via celery
+mccabe==0.5.3
 mock==2.0.0
-pydocstyle==1.0.0
+pbr==1.10.0               # via mock
+pep8-naming==0.4.1
 pep8==1.7.0
-pep8-naming==0.3.3
-pre-commit==0.8.0
-py==1.4.31
-pytest==2.9.1
+py==1.4.32                # via pytest
+pycodestyle==2.2.0        # via flake8
+pydocstyle==1.1.1
+pyflakes==1.3.0           # via flake8
+pytest-django==3.1.2
+pytest==3.0.5
+pytz==2016.10             # via celery
+six==1.10.0               # via mock
+vine==1.1.3               # via amqp
