+++ setup.cfg
@@ -8,7 +8,7 @@ max-line-length = 119
 max-complexity = 10
 statistics = true
 show-source = true
-exclude = */migrations/*,docs/*,env/*
+exclude = */migrations/*,docs/*,env/*,.tox/*
 
 [pydocstyle]
 add-ignore = D100,D101,D102,D103,D104,D105
@@ -16,8 +16,6 @@ match-dir = (?!tests|env|\.).*
 
 [coverage:run]
 source = .
-plugins =
-    django_coverage_plugin
 omit =
   */migrations/*
   */tests/*
