+++ setup.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 import os
 
-from setuptools import setup
+from setuptools import find_packages, setup
 
 from health_check import __version__
 
@@ -26,16 +26,7 @@ setup(
     license="BSD",
     keywords="django health check monitoring",
     url="https://github.com/KristianOellegaard/django-health-check",
-    packages=[
-        'health_check',
-        'health_check_celery',
-        'health_check_db',
-        'health_check_db.migrations',
-        'health_check_cache',
-        'health_check.backends',
-        'health_check_storage',
-        'health_check_storage_s3',
-    ],
+    packages=find_packages(exclude=['tests', 'docs']),
     long_description=read('README.rst'),
     classifiers=[
         'Topic :: Utilities',
