+++ tests/test_cache.py
@@ -1,13 +1,12 @@
 # -*- coding: utf-8 -*-
 from django.core.cache.backends.base import BaseCache, CacheKeyWarning
 from django.test import TestCase
-
 from mock import patch
 
 from health_check.backends.base import (
     ServiceReturnedUnexpectedResult, ServiceUnavailable
 )
-from health_check_cache.plugin_health_check import CacheBackend
+from health_check.cache.plugin_health_check import CacheBackend
 
 
 # A Mock version of the cache to use for testing
@@ -50,27 +49,27 @@ class HealthCheckCacheTests(TestCase):
     Ensures check_status returns/raises the expected result when the cache works, fails, or raises exceptions.
     """
 
-    @patch("health_check_cache.plugin_health_check.cache", MockCache())
+    @patch("health_check.cache.plugin_health_check.cache", MockCache())
     def test_check_status_working(self):
         cache_backend = CacheBackend()
         self.assertTrue(cache_backend.check_status())
 
     # check_status should raise ServiceUnavailable when values at cache key do not match
-    @patch("health_check_cache.plugin_health_check.cache", MockCache(set_works=False))
+    @patch("health_check.cache.plugin_health_check.cache", MockCache(set_works=False))
     def test_set_fails(self):
         cache_backend = CacheBackend()
         with self.assertRaises(ServiceUnavailable):
             cache_backend.check_status()
 
     # check_status should catch generic exceptions raised by set and convert to ServiceUnavailable
-    @patch("health_check_cache.plugin_health_check.cache", MockCache(set_raises=Exception))
+    @patch("health_check.cache.plugin_health_check.cache", MockCache(set_raises=Exception))
     def test_set_raises_generic(self):
         cache_backend = CacheBackend()
         with self.assertRaises(ServiceUnavailable):
             cache_backend.check_status()
 
     # check_status should catch CacheKeyWarning and convert to ServiceReturnedUnexpectedResult
-    @patch("health_check_cache.plugin_health_check.cache", MockCache(set_raises=CacheKeyWarning))
+    @patch("health_check.cache.plugin_health_check.cache", MockCache(set_raises=CacheKeyWarning))
     def test_set_raises_cache_key_warning(self):
         cache_backend = CacheBackend()
         with self.assertRaises(ServiceReturnedUnexpectedResult):
