+++ tests/testapp/__init__.py
@@ -1,3 +1,1 @@
 # -*- coding: utf-8 -*-
-import django
-django.setup()
