+++ tox.ini
@@ -2,6 +2,8 @@
 envlist = py{27,35,36}-dj{18,19,master},qa
 
 [testenv]
+setenv=
+    PYTHONPATH = {toxinidir}:$PYTHONPATH
 deps=
     -rrequirements-dev.txt
     dj18: https://github.com/django/django/archive/stable/1.8.x.tar.gz#egg=django
@@ -9,7 +11,7 @@ deps=
     dj110: https://github.com/django/django/archive/stable/1.10.x.tar.gz#egg=django
     djmaster: https://github.com/django/django/archive/master.tar.gz#egg=django
 commands=
-    py.test \
+    coverage run --source='health_check' -m 'pytest' \
         --basetemp={envtmpdir} \
         --ignore=.tox \
         {posargs}
